Move global classes to GraphQL directory
